+++ seedfarmer/output_utils.py
@@ -14,7 +14,6 @@
 
 from typing import Any, Dict, List, Optional
 
-import pyshorteners
 from rich.console import Console
 from rich.table import Table
 
@@ -142,30 +141,16 @@ def print_errored_modules(
     Parameters
     ----------
     description : str
-        The custom text to head the output table
+        The custom text 
     modules_data : List[Optional[ModuleDeploymentResponse]]
         The object containing the metadata related to the failure
     color : str, optional
-        The color of the text to head the description of the table, by default "red"
+        The color of the description text, by default "red"
     """
-
-    table = Table(title=f"[bold {color}]{description}", title_justify="left")
-
-    table.add_column("Deployemnt", justify="left", style="cyan", no_wrap=True)
-    table.add_column("Group", justify="left", style="magenta")
-    table.add_column("Module", justify="left", style="green")
-    table.add_column("CodeBuild Id", justify="left", style="white")
-    table.add_column("CodeBuild URL", justify="left", style="white")
+    console.print(f"[bold {color}]{description}", crop=False)
     for r_obj in modules_data:
         if r_obj and r_obj.status in ["ERROR", "error", "Error"]:
-            if r_obj.codeseeder_metadata:
-                url = r_obj.codeseeder_metadata.build_url
-                shortened_url = pyshorteners.Shortener().tinyurl.short(url) if url not in ["NA", "N/A"] else url
-                table.add_row(
-                    r_obj.deployment if r_obj and r_obj.deployment else None,
-                    r_obj.group if r_obj and r_obj.group else None,
-                    r_obj.module if r_obj and r_obj.module else None,
-                    r_obj.codeseeder_metadata.codebuild_build_id,
-                    shortened_url,
-                )
-    console.print(table)
+            console.print(f" [cyan]{r_obj.deployment}-{r_obj.group}-{r_obj.module}", crop=False)
+            console.print(
+                f"    {r_obj.codeseeder_metadata.build_url}", crop=False
+            ) if r_obj.codeseeder_metadata and r_obj.codeseeder_metadata.build_url else None